Правки в Program.cs Изменения визуала формы регистрации
