+++ ModulebankProject/Program.cs
@@ -1,8 +1,6 @@
 using FluentValidation;
 using MediatR;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.IdentityModel.Tokens;
-using Microsoft.OpenApi.Models;
+using ModulebankProject.Extensions;
 using ModulebankProject.Features.Accounts;
 using ModulebankProject.Features.Transactions;
 using ModulebankProject.Infrastructure.Data;
@@ -29,55 +27,7 @@ namespace ModulebankProject
             });
 
             builder.Services.AddControllers();
-            builder.Services.AddOpenApi();
-            builder.Services.AddSwaggerGen(options =>
-            {
-                var basePath = AppContext.BaseDirectory;
-
-                var xmlPath = Path.Combine(basePath, "ModulebankProject.xml");
-                options.UseInlineDefinitionsForEnums();
-                options.IncludeXmlComments(xmlPath);
-
-                options.CustomSchemaIds(id => id.FullName!.Replace('+','-'));
-
-                options.AddSecurityDefinition("Keycloak", new OpenApiSecurityScheme
-                {
-                    Type = SecuritySchemeType.OAuth2,
-                    Flows = new OpenApiOAuthFlows
-                    {
-                        Implicit = new OpenApiOAuthFlow
-                        {
-                            //AuthorizationUrl = new Uri(builder.Configuration["Keycloak:AuthUrl"]!),
-                            AuthorizationUrl = new Uri($"{builder.Configuration["Keycloak:AuthUrl"]}?client_id=client&response_type=token"),
-                            
-                            /*Scopes = new Dictionary<string, string>
-                            {
-                                {"openid", "openid"},
-                                {"profile", "profile"}
-                            }*/
-                        }
-                    }
-                });
-
-                var securityRequirement = new OpenApiSecurityRequirement
-                {
-                    {
-                        new OpenApiSecurityScheme
-                        {
-                            Reference = new OpenApiReference
-                            {
-                                Id = "Keycloak",
-                                Type = ReferenceType.SecurityScheme
-                            },
-                            In = ParameterLocation.Header,
-                            Name = "Bearer",
-                            Scheme = "Bearer"
-                        },
-                        []
-                    }
-                };
-                options.AddSecurityRequirement(securityRequirement);
-            });
+            builder.Services.AddCustomSwagger(builder.Configuration);
 
             builder.Services.AddMediatR(cfg => 
                 cfg.RegisterServicesFromAssembly(typeof(Program).Assembly));
@@ -96,17 +46,7 @@ namespace ModulebankProject
 
             //
             builder.Services.AddAuthorization();
-            builder.Services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-                .AddJwtBearer(o =>
-                {
-                    o.RequireHttpsMetadata = false;
-                    o.Audience = builder.Configuration["Authentication:Audience"];
-                    o.MetadataAddress = builder.Configuration["Authentication:MetadataAddress"]!;
-                    o.TokenValidationParameters = new TokenValidationParameters
-                    {
-                        ValidIssuer = builder.Configuration["Authentication:ValidIssuer"],
-                    };
-                });
+            builder.Services.AddAuth(builder.Configuration);
 
             var app = builder.Build();
 
@@ -115,9 +55,7 @@ namespace ModulebankProject
             //if (app.Environment.IsDevelopment())
             //{
                 app.UseSwagger();
-                app.UseSwaggerUI(c => { 
-                    c.OAuthClientId("client");
-                });
+                app.UseSwaggerUI();
             //}
             app.UseMiddleware<ErrorHandlingMiddleware>();
             app.UseHttpsRedirection();
